requirements.txt
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
numpy>=1.24.0
tqdm>=4.65.0
pdfplumber>=0.9.0
python-docx>=0.8.11
pytest>=7.4.0
pytest-cov>=4.1.0
